Sending an empty body list only once
If we don't send an empty body list the first time we are left with no
tracked bodies, chances are the client will keep the list of known bodies
cached as if it were still valid. We need to update it at least once with
an empty list. Tracking the last body count sent.